WIP: QIcon/QPixmap HiDPI support.
(second attempt, now with passing tests)

The reasoning behind this change is based on the observation
that producers and consumers of pixmaps are often 
far apart (for example QIcon::pixmap() and QLabel).
This means it makes sense to:
1) Modify producers to create 2X pixmaps on HiDPI displays
2) Carry the scale information with the pixmap
3) Modify consumers to account for the scale when needed.

Users of QPixmap now have to be aware of the DPI/scale
when using the pixmap size for layout calculations.
This includes application code so we probably need
to add a global HiDPI-aware switch if we go ahead
with this approach.

Make QIcon::pixmap() produce 2X pixmaps when running
on HiDPI screens. (ab)use the QPaintDevice dpi metrics
to tag the pixmap with the scale factor as a difference
between logical and physical DPI (not sure about this
but hey we are at least not adding new API).

Modify the "consumer" code paths necessary to make 
examples/mainwindows/application work. (See qpainter.cpp, 
qstyle.cpp, qlabel.cpp changes).

TODO: preserve dpi/scale across pixmap->image->pixmap
conversions.

Change-Id: Iaa0a898dc87701becfc4fe5688d76d3893866eee